+++ eras/dijkstra/impl/test/Main.hs
@@ -3,13 +3,22 @@
 
 module Main where
 
+import Cardano.Ledger.Block (Block (Block))
 import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Rules ()
 import Cardano.Ledger.Plutus (SLanguage (..))
+import Cardano.Protocol.Crypto (StandardCrypto)
+import qualified Cardano.Protocol.Leios.BlockHeader as Leios
+import qualified Test.Cardano.Base.QuickCheck as BaseQC
 import Test.Cardano.Ledger.Babbage.TxInfoSpec (txInfoSpec)
 import qualified Test.Cardano.Ledger.Babbage.TxInfoSpec as BabbageTxInfo
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Conway.Binary.RoundTrip (roundTripConwayCommonSpec)
+import Test.Cardano.Ledger.Core.Binary.RoundTrip (
+  roundTripAnnEraExpectation,
+  roundTripEraExpectation,
+ )
+import Test.Cardano.Ledger.Dijkstra.Arbitrary (genSmallDijkstraTxsBlockBody)
 import Test.Cardano.Ledger.Dijkstra.Binary.Annotator ()
 import qualified Test.Cardano.Ledger.Dijkstra.Binary.CddlSpec as Cddl
 import qualified Test.Cardano.Ledger.Dijkstra.Binary.Golden as GoldenBinary
@@ -29,6 +38,13 @@ main =
   ledgerEraTestMain @DijkstraEra $ do
     describe "RoundTrip" $ do
       roundTripConwayCommonSpec @DijkstraEra
+      prop "Block (Leios.Header)" $
+        BaseQC.withNumTests 25 $
+          forAll (Block <$> arbitrary <*> genSmallDijkstraTxsBlockBody) $ \block ->
+            conjoin
+              [ roundTripEraExpectation @DijkstraEra @(Block (Leios.Header StandardCrypto) DijkstraEra) block
+              , roundTripAnnEraExpectation @DijkstraEra @(Block (Leios.Header StandardCrypto) DijkstraEra) block
+              ]
     Cddl.spec
     GoldenSpec.spec
     roundTripJsonShelleyEraSpec @DijkstraEra